Export window.geolonia contents as geolonia when requiring as a NPM module
